feat: create_parents for add operations
Fixes: #25

Adds an option to the module which will create any parent objects/arrays
needed to reach the child member in an "add" operation. If the next
reference token in the path is either "0" or "-", an array will be
created. If it is any other number, the parent will not be created and
the operation will fail to avoid edge cases and quirks in idempotency.
In all other cases, an object is created.

This is in direct contradiction to Section 4.1 of RFC6902, which is why
this feature is OFF by default. However, it makes this module much more
useful when editing JSON configs.

Co-authored-by: davidolrik <david@olrik.dk>